split tenant.go to tenant.go + tenant_utils.go
